# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='300px' src='https://nitslaszlo.github.io/static/session/db_diagram.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedlik.cyclic.app"
        }
    ],
    "paths": {
...
        "/recipes": {
            "get": {
                "tags": ["recipes"],
                "summary": "Összes recept lekérdezése.",
                "description": "Az összes recept adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/user"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["recipes"],
                "summary": "Új recept mentése.",
                "description": "Az új recept mentése a 'hirdetesek' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/user"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett recepttel tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/recipes/{offset}/{limit}/{field}": {
            "get": {
                "tags": ["recipes"],
                "summary": "Paginált adatok lekérése.",
                "description": "receptek lekérdezése lapozáshoz.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "Az összes rekord száma.",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/user"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recipes/{offset}/{limit}/{field}/{filter}": {
            "get": {
                "tags": ["recipes"],
                "summary": "Paginált adatok lekérése szűréssel.",
                "description": "receptek lekérdezése lapozáshoz szűréssel.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "0"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "2"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    },
                    {
                        "name": "filter",
                        "description": "Szűrő - a 'cim' és a 'leiras' mezőkben keressük.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "sport"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "A szűrőnek megfelelő rekordok száma",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú, szűrőnek megfelelő rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/user"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "address": {
                "type": "object",
                "properties": {
                    "street": {
                        "type": "string",
                        "description": "A közterület neve és száma.",
                        "example": "Szent István út 7."
                    },
                    "city": {
                        "type": "string",
                        "description": "A település, ahol felhasználó lakik.",
                        "example": "Győr"
                    },
                    "country": {
                        "type": "string",
                        "description": "Az ország, ahol a felhasználó él.",
                        "example": "Hungary"
                    }
                }
            },
            "user": {
                "type": "object",
                "required": ["name", "email", "email_verified", "auto_login", "picture", "roles"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A felhasználó azonosítója (ObjectID).",
                        "example": "61b5f0c1f39e4edcf5b8a3d8"
                    },
                    "name": {
                        "type": "string",
                        "description": "A felhasználó neve",
                        "example": "student005"
                    },
                    "email": {
                        "type": "string",
                        "description": "A felhasználó által használt e-mail cím a regisztráció során",
                        "example": "student005@jedlik.eu"
                    },
                    "email_verified": {
                        "type": "boolean",
                        "description": "Igaz értékű, ha megadott e-mail-cím ellenőrzésre kerrült.",
                        "example": true
                    },
                    "auto_login": {
                        "type": "boolean",
                        "description": "Igaz értékű, ha a felhasználónál be van kapcsolva az automatikus bejelentkezés.",
                        "example": true
                    },
                    "picture": {
                        "type": "string",
                        "description": "A felhasználó avatarja/fényképe.",
                        "example": "https://lh3.googleusercontent.com/a/ALm5wu1tqlumAaDE2djCuUE2s4ubsOs6HOkdwfU6ftci=s96-c"
                    },
                    "password": {
                        "type": "string",
                        "description": "A felhasználó jelszavához tartozó hash-kód.",
                        "example": "$2b$10$HBH36EkxhwwOcI7UM/Km3uHUcgwPJjleZYS9fqOgjIr.qmvTUlGPu"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "A szerepköröket tartalmazó tömb, mely szepköket megkapott a felhasználó.",
                        "example": "[\"user\", \"admin\"]"
                    },
                    "address": {
                        "type": "object",
                        "$ref": "#/components/schemas/address",
                        "description": "A felhasználó lakcíme.",
                        "example": true
                    }
                }
            },
...
        }
    }
}